sends grayscale pictures successfully

+++ CommunicationLibrary/Messages/SharedObjects/PictureManager.py
@@ -7,7 +7,7 @@ class PictureManager:
     @staticmethod
     def splitPicture(picture, sizeOfParts):
         rows = picture.shape[0]
-        numberOfParts = math.ceil(rows/sizeOfParts)
+        numberOfParts = int(math.ceil(rows/sizeOfParts))
         splitFrames = np.array_split(picture, numberOfParts)
         return splitFrames, numberOfParts
